# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Agence
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /api
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /consultant:
    # our controller name
    x-swagger-router-controller: consultant
    get:
      operationId: getAll
      description: get the consultant list
      # define the type of response for Success "200" and Error
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/DefaultListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /consultant/earnings:
    # our controller name
    x-swagger-router-controller: consultant
    post:
      operationId: getEarning
      description: get earnings by consultants 
      parameters:
        - name: consultants
          description: Get earnings by co_usuario
          in: body
          required: true
          schema:
            $ref: "#/definitions/CaoUser"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/DefaultListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse" 
  /consultant/average:
    # our controller name
    x-swagger-router-controller: consultant
    post:
      operationId: getAverageFixedCost
      description: get get average fixed cost
      parameters:
        - name: consultants
          description: Get earnings by co_usuario
          in: body
          required: true
          schema:
            $ref: "#/definitions/AverageFixedCost"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/DefaultListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /consultant/performance:
      # our controller name
      x-swagger-router-controller: consultant
      post:
        operationId: getPerformanceComercial
        description: get get performance comercial
        parameters:
          - name: consultants
            description: Get earnings by co_usuario
            in: body
            required: true
            schema:
              $ref: "#/definitions/AverageFixedCost"
        responses:
          "200":
            description: Success
            schema:
              $ref: "#/definitions/DefaultListResponse"
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  DefaultListResponse:
    required:
      - data
    properties:
      allSalario:
        type: integer
      data:
        type: array
        items: 
          type: object
  AverageFixedCost:
    type: object
    properties:
      consultants:
        type: array
        items: 
          type: string
        description: Id users
    required:
      - consultants
  CaoUser:
    type: object
    properties:
      consultants:
        type: array
        items: 
          type: string
        description: Id users
      monthStart:
        type: string
        description: Month
      monthEnd:
        type: string
        description: Month
    required:
      - consultants
      - monthStart
      - monthEnd
  GeneralResponse:
    type: object
    properties:
      data:
        type: object
        description: returns 1 if successful
    required:
      - data
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string